Refactor scripts into feature folders and add money system
Scripts were reorganized into feature-based folders for better project structure. Added a basic player money system: building now costs money, the player's money is tracked in GameManager, and the UI updates accordingly. Minor UI and material adjustments were also made in the scene and material files, Preview material now more transparent.

+++ Assets/Scripts/Building/SpawnableObject.cs
@@ -7,5 +7,5 @@ public class SpawnableObject : ScriptableObject
     public GameObject prefab;
     public Sprite icon;
     public string objectName;
-    public int cost;
+    public float cost;
 }
